# Add squatter detection