Fixes #5 by implementing [sholladay/squatter](https://github.com/sholladay/squatter) and logging a warning for squatters, exiting code zero. Previously, we would exit with the "is unavailable" error. This makes the CLI more useful, as I believe that most people are willing to ask for a package name if it is not _really_ being used.

The only downside I could think of is a theoretical scenario where someone were using this CLI in a script and wanted to distinguish between the squatter and non-squatter cases, they would now have to parse the output rather than using the exit code. We _could_ use a different exit code, which wouldn't be unheard of, though I think it would be more surprising than not in the common case. A squatter is effectively success, or at least very close to it.

Note that I did not put this behind a flag ([mentioned here](https://github.com/sindresorhus/npm-name-cli/issues/5#issuecomment-318777586)), mainly because...
 - I couldn't think of a good name that was short and succinct but also descriptive and intuitive
 - Personally, I would prefer this to be the default behavior and I think others will too

## Performance

It is not very noticeable.


```console
$ time node cli.js chalk foo unicorn-cake
✖ chalk is unavailable
⚠ foo is squatted
✔ unicorn-cake is available

real	0m0.846s
user	0m0.296s
sys	0m0.048s
```

 - Before: with [`8bfdb3c`](https://github.com/sindresorhus/npm-name-cli/commit/8bfdb3c0ff8ab34abf160a9690637dcd7057ddac), I see it vary between about 0.5s - 1.3s, commonly around 0.75s.
 - After: with [`2086794`](https://github.com/sholladay/npm-name-cli/commit/20867948f0346ead4d5be74ec92a23c2ab02e620), I see it vary between about 0.7s - 1.6s, commonly around 0.9s.

This is measured from a fairly typical 50mbps down / 10mbps up XFINITY connection in Boston, MA.

The variance between runs due to network congestion and other factors, with or without this PR, is much bigger than any overhead added by the new functionality. And that is without any performance tuning, which I haven't even attempted (some ideas in https://github.com/sindresorhus/npm-name-cli/issues/5#issuecomment-318830776, though).