@sindresorhus 2017-11-05
The implementation here looks great, but the `squatter` module has a lot of false positives. I tried it just with a random list of words, and more than half of them were false positives:

```
❯ npm-name foo bar lol fun food cat dog horse bird snake
✖ foo is unavailable
⚠ bar is squatted
⚠ lol is squatted
⚠ fun is squatted
✖ food is unavailable
⚠ cat is squatted
⚠ dog is squatted
✖ horse is unavailable
⚠ bird is squatted
⚠ snake is squatted
```

But these are not squatted:
https://www.npmjs.com/package/bar
https://www.npmjs.com/package/lol
https://www.npmjs.com/package/fun
https://www.npmjs.com/package/cat
https://www.npmjs.com/package/dog
https://www.npmjs.com/package/bird
https://www.npmjs.com/package/snake