@sholladay 2017-11-05
I will look into that. 👍 

I was already aware of a couple of false positives, which I should have mentioned. It is a heuristic, after all. But I'm sure it could use some improvement.

Looking at just a few of your examples, `bird` and `snake` seem like squatters to me due to inactivity, so our definition is probably different. I'm keeping an open mind about what the definition _should_ be, though. 

Could be fun to try to find better thresholds using [mind](https://github.com/stevenmiller888/mind).

A question worth asking is, would you prefer for the algorithm to have false negatives more than false positives? As mentioned in the README, the algorithm is designed such that...

> Packages are guilty until proven innocent.

I decided to do that because most of the signals available from npm tell us _good_ things about a package. In other words, there is no dislike button on npm, and that leads to a sort o bias that `squatter` has to compensate for. The way I did it, if the answer is wrong, it will almost always be a false positive where `squatter` thinks a package is a squatter when humans would say it isn't. This seems good to me because my goal was to determine when a name is worth disputing if you want it. Authors are sometimes willing to give up legitimate non-squatter packages that have very few users, so my experience leads me to believe that it's better to return `true` when in doubt.