@sindresorhus 2018-04-01
Alright, let's try it out. We can get feedback from users and improve the algorithm.