@sindresorhus 2018-04-01
I think it would be useful to use the https://api-docs.npms.io/ API. It has a very reliable quality score, based on many factors: https://github.com/npms-io/npms-analyzer/blob/master/docs/architecture.md