@sindresorhus 2018-04-01
> Could be fun to try to find better thresholds using mind.

That would be very cool!